requirements.txt
django-widget-tweaks
django-filter